requirements.txt
flask
flask-cors
python-dotenv
boto3
gunicorn